Created routing for App, created navbar, fixed cursor style for MyButton.
