+++ src/main.js
@@ -1,6 +1,7 @@
 import { createApp } from 'vue'
 import App from "./App"
 import components from "@/components/UI/index"
+import router from "@/router/router";
 
 const app = createApp(App)
 
@@ -8,4 +9,6 @@ components.forEach(component => {
     app.component(component.name, component)
 })
 
-app.mount('#app')
+app
+    .use(router)
+    .mount('#app')
