+++ src/pages/About.vue
@@ -1,0 +1,13 @@
+<template>
+    <h1>Данное приложение было создано в рамках курса по Vue3</h1>
+</template>
+
+<script>
+export default {
+  name: "About"
+}
+</script>
+
+<style scoped>
+
+</style>
