+++ src/pages/Main.vue
@@ -1,0 +1,13 @@
+<template>
+    <h1 class="">Welcome to test App!</h1>
+</template>
+
+<script>
+export default {
+  name: "Main"
+}
+</script>
+
+<style scoped>
+
+</style>
