+++ src/pages/PostPage.vue
@@ -1,0 +1,15 @@
+<template>
+    <div class="">
+        <h1>Страница поста с ID = {{ $route.params.id }}</h1>
+    </div>
+</template>
+
+<script>
+export default {
+  name: "PostPage"
+}
+</script>
+
+<style scoped>
+
+</style>
